extract ScriptModules into own folder
