+++ modules/effect-manager.d.ts
@@ -1,0 +1,7 @@
+import { Firebot } from "../index";
+
+export type EffectManager = {
+  registerEffect: <EffectModel>(
+    effectType: Firebot.EffectType<EffectModel>
+  ) => void;
+};
