+++ modules/event-manager.d.ts
@@ -1,0 +1,21 @@
+type EventSource = {
+  id: string;
+  name: string;
+  events: Array<{
+    id: string;
+    name: string;
+    description: string;
+    cached?: boolean;
+    manualMetadata?: Record<string, unknown>;
+  }>;
+};
+
+export type EventManager = {
+  registerEventSource: (eventSource: EventSource) => void;
+  triggerEvent: (
+    sourceId: string,
+    eventId: string,
+    meta: Record<string, unknown>,
+    isManual?: boolean
+  ) => void;
+};
