+++ modules/frontend-communicator.d.ts
@@ -1,0 +1,11 @@
+export type FrontendCommunicator = {
+  send(eventName: string, data: unknown): void;
+  on<ExpectedArgs extends Array<any> = [], ReturnPayload = void>(
+    eventName: string,
+    callback: (...args: ExpectedArgs[]) => ReturnPayload
+  ): void;
+  onAsync<ExpectedArgs extends Array<any> = [], ReturnPayload = void>(
+    eventName: string,
+    callback: (...args: ExpectedArgs[]) => Promise<ReturnPayload>
+  ): void;
+};
