+++ modules/logger.d.ts
@@ -1,0 +1,10 @@
+interface LeveledLogMethod {
+  (msg: string, ...meta: any[]): void;
+}
+
+export type Logger = {
+  debug: LeveledLogMethod;
+  info: LeveledLogMethod;
+  warn: LeveledLogMethod;
+  error: LeveledLogMethod;
+};
