+++ modules/twitch-chat.d.ts
@@ -1,0 +1,7 @@
+export type TwitchChat = {
+  sendChatMessage(
+    message: string,
+    whisperTarget?: string,
+    accountType?: "bot" | "streamer"
+  ): void;
+};
