+++ package.json
@@ -3,9 +3,15 @@
   "version": "5.27.0",
   "description": "Types for Firebot's Custom Scripts",
   "main": "",
-  "scripts": {},
+  "scripts": {
+    "prettier:check": "prettier --check '**/*.{json,ts}'",
+    "prettier:format": "prettier --write '**/*.{json,ts}'"
+  },
   "author": "ebiggz",
   "license": "ISC",
   "types": "index.d.ts",
-  "typeScriptVersion": "4.0.2"
+  "typeScriptVersion": "4.0.2",
+  "devDependencies": {
+    "prettier": "2.2.1"
+  }
 }
